@@ -9,7 +9,6 @@
   },
   "author": "Darren Torpey",
   "license": "ISC",
-  "dependencies": {},
   "devDependencies": {
     "axios": "^0.16.1",
     "babel-cli": "^6.18.0",
@@ -22,10 +21,13 @@
     "babel-preset-env": "^1.5.2",
     "babel-preset-latest": "^6.24.1",
     "es-module-loader": "^1.3.5",
-    "eslint": "^4.0.0",
+    "eslint": "4.13.1",
+    "eslint-config-prettier": "2.9.0",
+    "eslint-plugin-prettier": "2.3.1",
     "express": "^4.14.0",
     "lodash": "^4.17.4",
     "mo-js": "^0.288.1",
+    "prettier": "1.9.2",
     "rxjs": "^5.4.0",
     "waaclock": "^0.5.3",
     "webpack": "^2.6.1",
