@@ -1,41 +1,41 @@
-import _ from 'lodash';
-import { Timeline } from 'scheduling';
-import { getRandomScreenCoords } from 'positioning';
-import mojs from 'mo-js';
-import Pool c from 'pool';
+import _ from 'lodash'
+import { Timeline } from 'scheduling'
+import { getRandomScreenCoords } from 'positioning'
+import mojs from 'mo-js'
+import Pool from 'pool'
 
 // Base class for complex effects
 export class Effect {
   constructor() {
-    this._elements = [];
-    this.timeline = new mojs.Timeline();
+    this._elements = []
+    this.timeline = new mojs.Timeline()
   }
 
   registerElement(...elements) {
-    elements.forEach(el => this.timeline.add(el));
+    elements.forEach(el => this.timeline.add(el))
 
-    this._elements = this._elements.concat(elements);
+    this._elements = this._elements.concat(elements)
   }
 
   moveTo(coords) {
     this._elements.forEach(el => {
-      el.tune(coords);
-    });
-    return this;
+      el.tune(coords)
+    })
+    return this
   }
 
   regenerate() {
-    this._elements.forEach(el => el.generate());
-    return this;
+    this._elements.forEach(el => el.generate())
+    return this
   }
 
   replay() {
-    this.timeline.replay();
-    return this;
+    this.timeline.replay()
+    return this
   }
 
   get elements() {
-    return this._elements;
+    return this._elements
   }
 }
 
@@ -59,44 +59,44 @@ function blastWaveShape({
     radius: { [sizeStart]: size },
     duration: speed,
     opacity: { 1: 0 },
-  });
+  })
 }
 
 export class DoubleCircle extends Effect {
   constructor(opts = {}) {
-    super();
+    super()
 
-    this.registerElement(blastWaveShape(opts));
+    this.registerElement(blastWaveShape(opts))
   }
 }
 
 // Made of circles
 export class Starburst {
   constructor(opts = {}) {
-    this.__id = parseInt(new Date());
-    this.initSet(opts);
+    this.__id = parseInt(new Date())
+    this.initSet(opts)
   }
 
   initSet(opts) {
-    this.circle = new DoubleCircle(opts);
+    this.circle = new DoubleCircle(opts)
 
-    this.timeline = Timeline.set(this.circle.elements);
+    this.timeline = Timeline.set(this.circle.elements)
   }
 
   moveTo(coords) {
-    this.circle.moveTo(coords);
+    this.circle.moveTo(coords)
   }
 
   play() {
-    this.timeline.replay();
+    this.timeline.replay()
   }
 
   moveToCoords(coords) {
-    this.moveTo(coords);
+    this.moveTo(coords)
 
-    this.timeline.replay();
+    this.timeline.replay()
 
-    return this;
+    return this
   }
 }
 
@@ -110,86 +110,88 @@ const SWIRL_OPTS = {
   swirlFrequency: 'rand(2,4)',
   swirlSize: 'rand(6,14)',
   onPlaybackComplete() {
-    $(this.el).remove();
+    $(this.el).remove()
   },
-};
+}
 
 export class BubbleField extends Effect {
   constructor(opts = { y: 250 }) {
-    super();
+    super()
 
-    this.registerElement(...this.makeBursts(opts));
+    this.registerElement(...this.makeBursts(opts))
   }
 
   makeBursts(opts) {
     opts = {
       height: 150,
       ...opts,
-    };
+    }
 
     const xCoords = [1, 2, 3, 4, 5, 6, 7, 8].map(i => ({
       fill: '#6d89fd',
       duration: `rand(200, ${_.random(400, 1200)})`,
       x: 50 + i * 150,
       y: { [opts.y]: opts.y - opts.height },
-    }));
+    }))
 
-    return xCoords.map(props => new WaterBurst(props));
+    return xCoords.map(props => new WaterBurst(props))
   }
 
   play() {
-    this.replay();
+    this.replay()
   }
 }
 
 export class WaterBurst extends Effect {
   static play() {
-    const waterBurst = new WaterBurst({ fill: '#4d89fd' });
-    waterBurst.play();
+    const waterBurst = new WaterBurst({ fill: '#4d89fd' })
+    waterBurst.play()
   }
 
   constructor(opts) {
-    super();
+    super()
 
     this.swirl1 = new mojs.ShapeSwirl({
       ...SWIRL_OPTS,
       ...opts,
-    });
+    })
 
     this.swirl2 = new mojs.ShapeSwirl({
       ...SWIRL_OPTS,
       ...opts,
       direction: -1,
-    });
+    })
 
     this.swirl3 = new mojs.ShapeSwirl({
       ...SWIRL_OPTS,
       ...opts,
-    });
+    })
 
     this.swirl4 = new mojs.ShapeSwirl({
       ...SWIRL_OPTS,
       ...opts,
     }).then((...args) => {
-      console.log('args', args);
-    });
+      console.log('args', args)
+    })
 
-    this.registerElement(this.swirl1, this.swirl2, this.swirl3, this.swirl4);
+    this.registerElement(this.swirl1, this.swirl2, this.swirl3, this.swirl4)
 
     this.timeline = Timeline.set([
       this.swirl1,
       this.swirl2,
       this.swirl3,
       this.swirl4,
-    ]);
+    ])
   }
 
   play() {
-    var { x, y: y } = getRandomScreenCoords(0);
+    var { x, y: y } = getRandomScreenCoords(0)
     // y = window.innerHeight * .99;
-    y = window.innerHeight * 1.05;
+    y = window.innerHeight * 1.05
 
-    this.moveTo({ x, y: { [y]: y - 200 } }).regenerate().replay();
+    this.moveTo({ x, y: { [y]: y - 200 } })
+      .regenerate()
+      .replay()
   }
 }
 
@@ -199,7 +201,7 @@ window.pools = {
   greenBlast: new Pool(),
   orangeBlast: new Pool(),
   lightBlueBlast: new Pool(),
-};
+}
 
 /**
  * ~~~~~~~~~~~~~~
@@ -207,8 +209,8 @@ window.pools = {
  * ~~~~~~~~~~~~~~
  */
 for (let i = 0; i < 10; i++) {
-  const size = _.random(2, 8);
-  const pixelSize = size * 25;
+  const size = _.random(2, 8)
+  const pixelSize = size * 25
 
   pools.greenBlast.add(
     new Starburst({
@@ -217,32 +219,32 @@ for (let i = 0; i < 10; i++) {
       speed: 2000 - pixelSize,
       color: 'rgba(49, 233, 149, 0.5)',
     })
-  );
+  )
 
   pools.blueBlast.add(
     new Starburst({
       size: _.random(50, 200),
       color: 'rgba(49, 153, 249, 0.5)',
     })
-  );
+  )
 
   pools.orangeBlast.add(
     new Starburst({
       size: _.random(100, 150),
       color: 'rgba(249, 153, 49, 0.5)',
     })
-  );
+  )
 
   pools.lightBlueBlast.add(
     new Starburst({
       size: _.random(50, 200),
       color: 'rgba(95, 210, 251, 0.5)',
     })
-  );
+  )
 }
 
 for (let i = 0; i < 30; i++) {
-  pools.waterBurst.add(new WaterBurst({ fill: '#4d89fd' }));
+  pools.waterBurst.add(new WaterBurst({ fill: '#4d89fd' }))
 }
 
 /**
@@ -250,14 +252,14 @@ for (let i = 0; i < 30; i++) {
  * A blaster gets a random blast from the pool and fires it at a random location on the screen
  */
 function blaster(effectName) {
-  return () => pools[effectName].next().moveToCoords(getRandomScreenCoords());
+  return () => pools[effectName].next().moveToCoords(getRandomScreenCoords())
 }
 
 /**
  * Makes a player for the given effect
  */
 function player(effectName) {
-  return () => pools[effectName].next().play();
+  return () => pools[effectName].next().play()
 }
 
 const randomSplashEvents = [
@@ -265,13 +267,13 @@ const randomSplashEvents = [
   'blueBlast',
   'orangeBlast',
   'lightBlueBlast',
-];
+]
 
 const Effects = _.fromPairs([
   ['waterBurst', player('waterBurst')],
   ...randomSplashEvents.map(eff => [eff, blaster(eff)]),
-]);
+])
 
-window._Effects = Effects;
+window._Effects = Effects
 
-export { Effects };
+export { Effects }
