@@ -1,13 +1,13 @@
-import mojs from 'mo-js';
+import mojs from 'mo-js'
 
 // Mo.js utility
 export const Timeline = {
   set(...items) {
     if (items.length === 1) {
-      items = [items];
+      items = [items]
     }
-    const timeline = new mojs.Timeline();
-    timeline.add(...items);
-    return timeline;
+    const timeline = new mojs.Timeline()
+    timeline.add(...items)
+    return timeline
   },
-};
+}
